fix,dev

+++ src/artist/artist.controller.ts
@@ -3,13 +3,19 @@ import { ArtistService } from './artist.service';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
 import { validate } from 'uuid';
+import { validate as classValidate }  from 'class-validator';
 import { BadRequestException, ForbiddenException, NotFoundException } from '@nestjs/common/exceptions';
 @Controller('artist')
 export class ArtistController {
   constructor(private readonly artistService: ArtistService) {}
 
   @Post()
-  create(@Body() createArtistDto: CreateArtistDto) {
+  async create(@Body() createArtistDto: CreateArtistDto) {
+    const errors =await classValidate(createArtistDto)
+
+    if(errors.length > 0) {
+      throw new BadRequestException(errors.toString());
+    }
     return this.artistService.create(createArtistDto);
   }
 
@@ -53,6 +59,13 @@ export class ArtistController {
     if(!validate(id)) {
       throw new BadRequestException('Invalid artist id'); ;
     }
+
+    const artist = this.artistService.findOne(id)
+
+    if(!artist){
+      throw new NotFoundException('Artist not found');
+    }
+
     return this.artistService.remove(id);
   }
 }
